fix: Addressing PR comments for splitting the new classes, structs, and enums in the TestRunner.Run tool out into seperate files in a new TestRunner subfolder. Also fixing the test validation matching for assembly and class input names.
