+++ Assets/root/Editor/Scripts/API/Tool/TestRunner.Run.cs
@@ -3,13 +3,13 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using com.IvanMurzak.Unity.MCP.Common;
 using com.IvanMurzak.ReflectorNet.Utils;
 using UnityEditor.TestTools.TestRunner.Api;
 using UnityEngine;
+using com.IvanMurzak.Unity.MCP.Editor.API.TestRunner;
 
 namespace com.IvanMurzak.Unity.MCP.Editor.API
 {
@@ -103,7 +103,7 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
 
                     Debug.Log($"[TestRunner] Running {testMode} tests.");
                     var resultCollector = await RunSingleTestModeWithCollector(testModeEnum, testRunnerApi, filterParams, timeoutMs);
-                    return FormatTestResults(resultCollector);
+                    return resultCollector.FormatTestResults();
                 }
             }
             catch (OperationCanceledException)
@@ -196,13 +196,13 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
         /// <summary>
         /// Creates a regex pattern for class filtering that matches Unity's Filter.groupNames behavior.
         /// This ensures our validation logic (CountFilteredTests) matches exactly what Unity's TestRunner will execute.
-        /// Pattern: "^.*\.{className}$" - matches any test class with the specified name regardless of namespace.
+        /// Pattern: "^.*\.{className}\.[^\.]+$" - matches any test class with the specified name followed by a method name.
         /// </summary>
         /// <param name="className">The class name to filter by</param>
         /// <returns>Regex pattern for Unity's Filter.groupNames field</returns>
         private static string CreateClassRegexPattern(string className)
         {
-            return $"^.*\\.{EscapeRegex(className)}$";
+            return $"^.*\\.{EscapeRegex(className)}\\.[^\\.]+$";
         }
 
         /// <summary>
@@ -286,10 +286,18 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
             {
                 bool matches = false;
 
-                // Check assembly filter
-                if (!string.IsNullOrEmpty(filterParams.TestAssembly) && test.FullName.Contains(filterParams.TestAssembly))
+                // Check assembly filter using UniqueName which contains assembly information
+                if (!string.IsNullOrEmpty(filterParams.TestAssembly))
                 {
-                    matches = true;
+                    var dllIndex = test.UniqueName.IndexOf(".dll");
+                    if (dllIndex > 0)
+                    {
+                        var assemblyName = test.UniqueName.Substring(0, dllIndex);
+                        if (assemblyName.Equals(filterParams.TestAssembly, StringComparison.OrdinalIgnoreCase))
+                        {
+                            matches = true;
+                        }
+                    }
                 }
 
                 // Check namespace filter using same regex pattern as Filter.groupNames (ensures sync with Unity's execution)
@@ -346,7 +354,6 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
 
             try
             {
-                var editModeTestCount = 0;
                 var remainingTimeoutMs = timeoutMs;
 
                 // Run EditMode tests if they exist
@@ -357,7 +364,6 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
                     var editModeCollector = await RunSingleTestModeWithCollector(TestMode.EditMode, testRunnerApi, filterParams, timeoutMs);
                     combinedCollector.AddResults(editModeCollector);
 
-                    editModeTestCount = editModeCollector.GetSummary().TotalTests;
                     var editModeDuration = DateTime.Now - editModeStartTime;
                     remainingTimeoutMs = Math.Max(1000, timeoutMs - (int)editModeDuration.TotalMilliseconds);
 
@@ -372,7 +378,7 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
                 if (runPlayMode)
                 {
                     Debug.Log($"[TestRunner] Starting PlayMode tests with {remainingTimeoutMs}ms timeout...");
-                    var playModeCollector = await RunSingleTestModeWithCollector(TestMode.PlayMode, testRunnerApi, filterParams, remainingTimeoutMs, editModeTestCount);
+                    var playModeCollector = await RunSingleTestModeWithCollector(TestMode.PlayMode, testRunnerApi, filterParams, remainingTimeoutMs);
                     combinedCollector.AddResults(playModeCollector);
                     Debug.Log($"[TestRunner] PlayMode tests completed.");
                 }
@@ -388,7 +394,7 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
                 // Format combined results - handle case where only one mode ran
                 if (runEditMode && runPlayMode)
                 {
-                    return FormatCombinedResults(combinedCollector);
+                    return combinedCollector.FormatCombinedResults();
                 }
                 else
                 {
@@ -396,7 +402,7 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
                     var collectors = combinedCollector.GetAllCollectors();
                     if (collectors.Any())
                     {
-                        return FormatTestResults(collectors.First());
+                        return collectors.First().FormatTestResults();
                     }
                     else
                     {
@@ -410,7 +416,7 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
             }
         }
 
-        private async Task<TestResultCollector> RunSingleTestModeWithCollector(TestMode testMode, TestRunnerApi testRunnerApi, TestFilterParameters filterParams, int timeoutMs, int previousTestCount = 0)
+        private async Task<TestResultCollector> RunSingleTestModeWithCollector(TestMode testMode, TestRunnerApi testRunnerApi, TestFilterParameters filterParams, int timeoutMs)
         {
             var filter = CreateTestFilter(testMode, filterParams);
             var runNumber = testMode == TestMode.EditMode
@@ -442,366 +448,6 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
             }
         }
 
-        private static string FormatTestResults(TestResultCollector collector)
-        {
-            var results = collector.GetResults();
-            var summary = collector.GetSummary();
-            var logs = collector.GetLogs();
 
-            var output = new StringBuilder();
-            output.AppendLine("[Success] Test execution completed.");
-            output.AppendLine();
-
-            // Summary
-            output.AppendLine("=== TEST SUMMARY ===");
-            output.AppendLine($"Status: {summary.Status}");
-            output.AppendLine($"Total: {summary.TotalTests}");
-            output.AppendLine($"Passed: {summary.PassedTests}");
-            output.AppendLine($"Failed: {summary.FailedTests}");
-            output.AppendLine($"Skipped: {summary.SkippedTests}");
-            output.AppendLine($"Duration: {summary.Duration:hh\\:mm\\:ss\\.fff}");
-            output.AppendLine();
-
-            // Individual test results
-            if (results.Any())
-            {
-                output.AppendLine("=== TEST RESULTS ===");
-                foreach (var result in results)
-                {
-                    output.AppendLine($"[{result.Status}] {result.Name}");
-                    output.AppendLine($"  Duration: {result.Duration:ss\\.fff}s");
-
-                    if (!string.IsNullOrEmpty(result.Message))
-                    {
-                        output.AppendLine($"  Message: {result.Message}");
-                    }
-
-                    if (!string.IsNullOrEmpty(result.StackTrace))
-                    {
-                        output.AppendLine($"  Stack Trace: {result.StackTrace}");
-                    }
-                    output.AppendLine();
-                }
-            }
-
-            // Console logs
-            if (logs.Any())
-            {
-                output.AppendLine("=== CONSOLE LOGS ===");
-                foreach (var log in logs)
-                {
-                    output.AppendLine(log);
-                }
-            }
-
-            return output.ToString();
-        }
-
-        private static string FormatCombinedResults(CombinedTestResultCollector collector)
-        {
-            var results = collector.GetResults();
-            var summary = collector.GetSummary();
-            var editModeSummary = collector.GetEditModeSummary();
-            var playModeSummary = collector.GetPlayModeSummary();
-            var logs = collector.GetLogs();
-
-            var output = new StringBuilder();
-            output.AppendLine("[Success] Combined test execution completed.");
-            output.AppendLine();
-
-            // Combined Summary
-            output.AppendLine("=== COMBINED TEST SUMMARY ===");
-            var overallStatusColored = summary.Status == TestRunStatus.Passed
-                ? "<color=green>✅</color>"
-                : "<color=red>❌</color>";
-            output.AppendLine($"Overall Status: {summary.Status} {overallStatusColored}");
-            output.AppendLine($"Total Tests: {summary.TotalTests}");
-            output.AppendLine($"Total Passed: {summary.PassedTests}");
-            output.AppendLine($"Total Failed: {summary.FailedTests}");
-            output.AppendLine($"Total Skipped: {summary.SkippedTests}");
-            output.AppendLine($"Total Duration: {summary.Duration:hh\\:mm\\:ss\\.fff}");
-            output.AppendLine();
-
-            // EditMode Summary
-            if (editModeSummary.TotalTests > 0)
-            {
-                var editModeStatusColored = editModeSummary.Status == TestRunStatus.Passed
-                    ? "<color=green>✅</color>"
-                    : "<color=red>❌</color>";
-                output.AppendLine("=== EDITMODE TEST SUMMARY ===");
-                output.AppendLine($"Status: {editModeSummary.Status} {editModeStatusColored}");
-                output.AppendLine($"Total: {editModeSummary.TotalTests}");
-                output.AppendLine($"Passed: {editModeSummary.PassedTests}");
-                output.AppendLine($"Failed: {editModeSummary.FailedTests}");
-                output.AppendLine($"Skipped: {editModeSummary.SkippedTests}");
-                output.AppendLine($"Duration: {editModeSummary.Duration:hh\\:mm\\:ss\\.fff}");
-                output.AppendLine();
-            }
-
-            // PlayMode Summary
-            if (playModeSummary.TotalTests > 0)
-            {
-                var playModeStatusColored = playModeSummary.Status == TestRunStatus.Passed
-                    ? "<color=green>✅</color>"
-                    : "<color=red>❌</color>";
-                output.AppendLine("=== PLAYMODE TEST SUMMARY ===");
-                output.AppendLine($"Status: {playModeSummary.Status} {playModeStatusColored}");
-                output.AppendLine($"Total: {playModeSummary.TotalTests}");
-                output.AppendLine($"Passed: {playModeSummary.PassedTests}");
-                output.AppendLine($"Failed: {playModeSummary.FailedTests}");
-                output.AppendLine($"Skipped: {playModeSummary.SkippedTests}");
-                output.AppendLine($"Duration: {playModeSummary.Duration:hh\\:mm\\:ss\\.fff}");
-                output.AppendLine();
-            }
-
-            // Individual test results
-            if (results.Any())
-            {
-                output.AppendLine("=== TEST RESULTS ===");
-                foreach (var result in results)
-                {
-                    output.AppendLine($"[{result.Status}] {result.Name}");
-                    output.AppendLine($"  Duration: {result.Duration:ss\\.fff}s");
-
-                    if (!string.IsNullOrEmpty(result.Message))
-                    {
-                        output.AppendLine($"  Message: {result.Message}");
-                    }
-
-                    if (!string.IsNullOrEmpty(result.StackTrace))
-                    {
-                        output.AppendLine($"  Stack Trace: {result.StackTrace}");
-                    }
-                    output.AppendLine();
-                }
-            }
-
-            // Console logs
-            if (logs.Any())
-            {
-                output.AppendLine("=== CONSOLE LOGS ===");
-                foreach (var log in logs)
-                {
-                    output.AppendLine(log);
-                }
-            }
-
-            return output.ToString();
-        }
-
-    }
-
-    // Helper classes for test result collection
-    public class TestResultCollector : ICallbacks
-    {
-        private readonly List<TestResultData> _results = new();
-        private readonly List<string> _logs = new();
-        private readonly TaskCompletionSource<bool> _completionSource = new();
-        private readonly TestSummaryData _summary = new();
-        private DateTime _startTime;
-        private readonly TestMode _testMode;
-        private readonly int _runNumber;
-
-        public TestResultCollector(TestMode testMode, int runNumber = 1)
-        {
-            _testMode = testMode;
-            _runNumber = runNumber;
-        }
-
-        public void RunStarted(ITestAdaptor testsToRun)
-        {
-            _startTime = DateTime.Now;
-            var testCount = CountTests(testsToRun);
-
-            _summary.TotalTests = testCount;
-            Debug.Log($"[TestRunner] Run {_runNumber} ({_testMode}) started: {testCount} tests.");
-        }
-
-        public void RunFinished(ITestResultAdaptor result)
-        {
-            var endTime = DateTime.Now;
-            var duration = endTime - _startTime;
-            _summary.Duration = duration;
-            if (_summary.FailedTests > 0)
-            {
-                _summary.Status = TestRunStatus.Failed;
-            }
-            else if (_summary.PassedTests > 0)
-            {
-                _summary.Status = TestRunStatus.Passed;
-            }
-            else
-            {
-                _summary.Status = TestRunStatus.Unknown;
-            }
-
-            Debug.Log($"[TestRunner] Run {_runNumber} ({_testMode}) finished with {CountTests(result.Test)} test results. Result status: {result.TestStatus}");
-            Debug.Log($"[TestRunner] Final duration: {duration:mm\\:ss\\.fff}. Completed: {_results.Count}/{_summary.TotalTests}");
-
-            _completionSource.TrySetResult(true);
-        }
-
-        public void TestStarted(ITestAdaptor test)
-        {
-            // Test started - could log this if needed
-        }
-
-        public void TestFinished(ITestResultAdaptor result)
-        {
-            // Only count actual tests, not test suites
-            if (!result.Test.IsSuite)
-            {
-                var testResult = new TestResultData
-                {
-                    Name = result.Test.FullName,
-                    Status = result.TestStatus.ToString(),
-                    Duration = TimeSpan.FromSeconds(result.Duration),
-                    Message = result.Message,
-                    StackTrace = result.StackTrace
-                };
-
-                _results.Add(testResult);
-
-                var statusEmoji = result.TestStatus switch
-                {
-                    TestStatus.Passed => "<color=green>✅</color>",
-                    TestStatus.Failed => "<color=red>❌</color>",
-                    TestStatus.Skipped => "<color=yellow>⚠️</color>",
-                    _ => ""
-                };
-
-                Debug.Log($"[TestRunner] {statusEmoji} Test finished: {result.Test.FullName} - {result.TestStatus} ({_results.Count}/{_summary.TotalTests})");
-
-                // Update summary counts
-                switch (result.TestStatus)
-                {
-                    case TestStatus.Passed:
-                        _summary.PassedTests++;
-                        break;
-                    case TestStatus.Failed:
-                        _summary.FailedTests++;
-                        break;
-                    case TestStatus.Skipped:
-                        _summary.SkippedTests++;
-                        break;
-                }
-
-                // Update duration as tests complete
-                _summary.Duration = DateTime.Now - _startTime;
-
-                // Check if all tests are complete
-                if (_results.Count >= _summary.TotalTests)
-                {
-                    Debug.Log($"[TestRunner] All tests completed via TestFinished. Final duration: {_summary.Duration:mm\\:ss\\.fff}");
-                    _completionSource.TrySetResult(true);
-                }
-            }
-        }
-
-        public async Task WaitForCompletionAsync(CancellationToken cancellationToken)
-        {
-            var tcs = new TaskCompletionSource<bool>();
-            using (cancellationToken.Register(() => tcs.TrySetCanceled()))
-            {
-                var completedTask = await Task.WhenAny(_completionSource.Task, tcs.Task);
-                if (completedTask == tcs.Task)
-                {
-                    cancellationToken.ThrowIfCancellationRequested();
-                }
-                await _completionSource.Task; // Re-await to get the result or exception
-            }
-        }
-
-        public List<TestResultData> GetResults() => _results;
-        public TestSummaryData GetSummary() => _summary;
-        public List<string> GetLogs() => _logs;
-        public TestMode GetTestMode() => _testMode;
-
-        public static int CountTests(ITestAdaptor test)
-        {
-            if (!test.HasChildren)
-                return test.IsSuite
-                    ? 0
-                    : 1;
-
-            return test.Children.Sum(CountTests);
-        }
-    }
-
-    public class CombinedTestResultCollector
-    {
-        private readonly List<TestResultData> _allResults = new();
-        private readonly List<string> _allLogs = new();
-        private readonly List<TestResultCollector> _collectors = new();
-        private TestSummaryData _combinedSummary = new();
-        private TestSummaryData _editModeSummary = new();
-        private TestSummaryData _playModeSummary = new();
-
-        public void AddResults(TestResultCollector collector)
-        {
-            var results = collector.GetResults();
-            var summary = collector.GetSummary();
-            var logs = collector.GetLogs();
-            var testMode = collector.GetTestMode();
-
-            _collectors.Add(collector);
-            _allResults.AddRange(results);
-            _allLogs.AddRange(logs);
-
-            _combinedSummary.TotalTests += summary.TotalTests;
-            _combinedSummary.PassedTests += summary.PassedTests;
-            _combinedSummary.FailedTests += summary.FailedTests;
-            _combinedSummary.SkippedTests += summary.SkippedTests;
-
-            if (testMode == TestMode.EditMode)
-            {
-                _editModeSummary = summary;
-            }
-            else if (testMode == TestMode.PlayMode)
-            {
-                _playModeSummary = summary;
-            }
-        }
-
-        public void SetTotalDuration(TimeSpan duration)
-        {
-            _combinedSummary.Duration = duration;
-            _combinedSummary.Status = _combinedSummary.FailedTests > 0
-                ? TestRunStatus.Failed
-                : TestRunStatus.Passed;
-        }
-
-        public List<TestResultData> GetResults() => _allResults;
-        public TestSummaryData GetSummary() => _combinedSummary;
-        public TestSummaryData GetEditModeSummary() => _editModeSummary;
-        public TestSummaryData GetPlayModeSummary() => _playModeSummary;
-        public List<string> GetLogs() => _allLogs;
-        public List<TestResultCollector> GetAllCollectors() => _collectors;
-    }
-
-    public class TestResultData
-    {
-        public string Name { get; set; } = string.Empty;
-        public string Status { get; set; } = string.Empty;
-        public TimeSpan Duration { get; set; }
-        public string? Message { get; set; }
-        public string? StackTrace { get; set; }
-    }
-
-    public enum TestRunStatus
-    {
-        Unknown,
-        Passed,
-        Failed
-    }
-
-    public class TestSummaryData
-    {
-        public TestRunStatus Status { get; set; } = TestRunStatus.Unknown;
-        public int TotalTests { get; set; }
-        public int PassedTests { get; set; }
-        public int FailedTests { get; set; }
-        public int SkippedTests { get; set; }
-        public TimeSpan Duration { get; set; }
     }
 }
