+++ Assets/root/Editor/Scripts/API/Tool/TestRunner/CombinedTestResultCollector.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fff8a3c10f9baa745b65ed0ce595ef92
