+++ Assets/root/Editor/Scripts/API/Tool/TestRunner/TestResultCollector.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f8b94719feeb71649aa6912965b85137
