+++ Assets/root/Editor/Scripts/API/Tool/TestRunner/TestRunStatus.cs
@@ -1,0 +1,11 @@
+#pragma warning disable CS8632 // The annotation for nullable reference types should only be used in code within a '#nullable' annotations context.
+
+namespace com.IvanMurzak.Unity.MCP.Editor.API.TestRunner
+{
+    public enum TestRunStatus
+    {
+        Unknown,
+        Passed,
+        Failed
+    }
+}
